requirements.txt
pymodbus>=3.11.4
flask>=3.1.2
pyyaml>=6.0.3
cryptography>=46.0.3
werkzeug>=3.1.5
waitress>=3.0.2
paho-mqtt>=2.1.0
requests>=2.32.5
pytest>=8.0.0
ruff>=0.1.0
flasgger>=0.9.7
flask-limiter>=3.5.0
